requirements.txt
click==8.1.3
colorama==0.4.6
coverage==7.1.0
dnspython==2.3.0
Faker==16.7.0
Flask==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
Werkzeug==2.2.2
pytest==7.2.1
pytest-cov==4.0.0
